Multiple features, single time step -> single feature, single time step
Presentation of  a set of recurrent neural networks architectures for one-to-one sequence problems, including variety of the available secondary options such as TensorBoard output and peephole connections models. The prediction is based on  multiple features (independent variables) at a single time step, while the outcome (target) is a single feature at a single time step.  The architectures assume regression problem (the outcome is assumed to be continuous), but can be easily adapted for classification problems.